[Change] Made Map keys 64 bit. Added DynNNets.

+++ StockTrader/Include/JLib/Map.h
@@ -1,5 +1,6 @@
 #pragma once
 #include <cstdint>
+#include <JLib/KeyPair.h>
 
 namespace jv
 {
@@ -8,17 +9,17 @@ namespace jv
 	struct Map final
 	{
 		KeyPair<T>* ptr = nullptr;
-		uint32_t length = 0;
+		uint64_t length = 0;
 		uint32_t count = 0;
 
-		void Insert(const T& instance, uint32_t key);
-		[[nodiscard]] T* Contains(uint32_t key) const;
+		void Insert(const T& instance, uint64_t key);
+		[[nodiscard]] T* Contains(uint64_t key) const;
 		void Erase(T& value);
-		[[nodiscard]] uint32_t Hash(uint32_t key) const;
+		[[nodiscard]] uint64_t Hash(uint64_t key) const;
 	};
 
 	template <typename T>
-	void Map<T>::Insert(const T& instance, const uint32_t key)
+	void Map<T>::Insert(const T& instance, const uint64_t key)
 	{
 		assert(count < length);
 
@@ -26,11 +27,11 @@ namespace jv
 		if (Contains(key))
 			return;
 
-		const uint32_t hash = Hash(key);
+		const uint64_t hash = Hash(key);
 
-		for (uint32_t i = 0; i < length; ++i)
+		for (uint64_t i = 0; i < length; ++i)
 		{
-			const uint32_t index = (hash + i) % length;
+			const uint64_t index = (hash + i) % length;
 			auto& keyPair = ptr[index];
 			// Set to true the first time the key group has been found.
 			if (keyPair.key != SIZE_MAX)
@@ -44,16 +45,16 @@ namespace jv
 	}
 
 	template <typename T>
-	T* Map<T>::Contains(const uint32_t key) const
+	T* Map<T>::Contains(const uint64_t key) const
 	{
 		assert(count <= length);
 
 		// Get and use the hash as an index.
-		const uint32_t hash = Hash(key);
+		const uint64_t hash = Hash(key);
 
-		for (uint32_t i = 0; i < length; ++i)
+		for (uint64_t i = 0; i < length; ++i)
 		{
-			const uint32_t index = (hash + i) % length;
+			const uint64_t index = (hash + i) % length;
 			auto& keyPair = ptr[index];
 
 			// If the hash is different, continue.
@@ -67,7 +68,7 @@ namespace jv
 	template <typename T>
 	void Map<T>::Erase(T& value)
 	{
-		uint32_t index;
+		uint64_t index;
 		const bool contains = Contains(value, index);
 		assert(contains);
 		assert(count > 0);
@@ -75,10 +76,10 @@ namespace jv
 		auto& keyPair = ptr[index];
 
 		// Check how big the key group is.
-		uint32_t i = 1;
+		uint64_t i = 1;
 		while (i < length)
 		{
-			const uint32_t otherIndex = (index + i) % length;
+			const uint64_t otherIndex = (index + i) % length;
 			auto& otherKeyPair = ptr[otherIndex];
 			if (otherKeyPair.key != keyPair.key)
 				break;
@@ -95,7 +96,7 @@ namespace jv
 	}
 
 	template <typename T>
-	uint32_t Map<T>::Hash(const uint32_t key) const
+	uint64_t Map<T>::Hash(const uint64_t key) const
 	{
 		return key % length;
 	}
